- moves startTransportBindings, stopTransportBindings, applyFunctionFromBinding to a dedicated file;
