+++ lib/iotagent-ul.js
@@ -25,70 +25,11 @@
 
 var iotAgentLib = require('iotagent-node-lib'),
     async = require('async'),
-    fs = require('fs'),
-    path = require('path'),
     context = {
         op: 'IoTA-UL20.Agent'
     },
     config = require('./configService'),
-    transportBindings = [];
-
-/**
- * Start all the transport protocol bindings found in the bindings directory.
- *
- * @param {Object} newConfig        Configuration object to start the bindings
- */
-function startTransportBindings(newConfig, callback) {
-    function invokeBinding(binding, callback) {
-        binding.start(callback);
-    }
-
-    var bindings = fs.readdirSync(path.join(__dirname, './bindings'));
-
-    transportBindings = bindings.map(function(item) {
-        return require('./bindings/' + item);
-    });
-
-    async.map(transportBindings, invokeBinding, callback);
-}
-
-/**
- * Stop all the transport protocol bindings of the agent.
- */
-function stopTransportBindings(callback) {
-    function invokeBinding(binding, callback) {
-        binding.stop(callback);
-    }
-
-    async.map(transportBindings, invokeBinding, callback);
-}
-
-/**
- * Execute the function given by the 'functionName' parameter for all the transport bindings, with the arguments
- * given in the 'argument' array. If the optional parameter protocol is not null, the function will only be
- * executed in the plugin of the selected protocol.
- *
- * @param {Array} argument          Array of arguments to call the function with.
- * @param {String} functionName     Name of the function to call in every transport plugin.
- * @param {String} protocol         Transport protocol where the function must be executed.
- */
-function applyFunctionFromBinding(argument, functionName, protocol, callback) {
-    config.getLogger().debug('Looking for bindings for the function [%s] and protocol [%s]', functionName, protocol);
-
-    function addHandler(current, binding) {
-        if (binding[functionName] && (!protocol || binding.protocol === protocol)) {
-            var args = [binding[functionName]].concat(argument),
-                boundFunction = binding[functionName].bind.apply(binding[functionName], args);
-
-            config.getLogger().debug('Binding found for function [%s] and protocol [%s]', functionName, protocol);
-            current.push(boundFunction);
-        }
-
-        return current;
-    }
-
-    async.series(transportBindings.reduce(addHandler, []), callback);
-}
+    transportSelector = require('./transportSelector');
 
 /**
  * Calls all the device provisioning handlers for each transport protocol binding whenever a new device is provisioned
@@ -97,7 +38,7 @@ function applyFunctionFromBinding(argument, functionName, protocol, callback) {
  * @param {Object} device           Device provisioning information.
  */
 function deviceProvisioningHandler(device, callback) {
-    applyFunctionFromBinding([device], 'deviceProvisioningHandler', null, function(error, devices) {
+    transportSelector.applyFunctionFromBinding([device], 'deviceProvisioningHandler', null, function(error, devices) {
         if (error) {
             callback(error);
         } else {
@@ -113,7 +54,7 @@ function deviceProvisioningHandler(device, callback) {
  * @param {Object} configuration     Configuration provisioning information.
  */
 function configurationHandler(configuration, callback) {
-    applyFunctionFromBinding([configuration], 'configurationHandler', null, callback);
+    transportSelector.applyFunctionFromBinding([configuration], 'configurationHandler', null, callback);
 }
 
 /**
@@ -137,7 +78,7 @@ function commandHandler(id, type, service, subservice, attributes, callback) {
             );
             callback(error);
         } else {
-            applyFunctionFromBinding(
+            transportSelector.applyFunctionFromBinding(
                 [device, attributes],
                 'commandHandler',
                 device.transport || config.getConfig().defaultTransport,
@@ -155,7 +96,7 @@ function commandHandler(id, type, service, subservice, attributes, callback) {
  * @param {Array} values                Values recieved in the notification.
  */
 function notificationHandler(device, values, callback) {
-    applyFunctionFromBinding(
+    transportSelector.applyFunctionFromBinding(
         [device, values],
         'notificationHandler',
         device.transport || config.getConfig().defaultTransport,
@@ -217,7 +158,7 @@ function start(newConfig, callback) {
 
             iotAgentLib.setNotificationHandler(notificationHandler);
 
-            startTransportBindings(newConfig, callback);
+            transportSelector.startTransportBindings(newConfig, callback);
         }
     });
 }
@@ -228,10 +169,13 @@ function start(newConfig, callback) {
  */
 function stop(callback) {
     config.getLogger().info(context, 'Stopping IoT Agent: ');
-    async.series([stopTransportBindings, iotAgentLib.resetMiddlewares, iotAgentLib.deactivate], function() {
-        config.getLogger().info('Agent stopped');
-        callback();
-    });
+    async.series(
+        [transportSelector.stopTransportBindings, iotAgentLib.resetMiddlewares, iotAgentLib.deactivate],
+        function() {
+            config.getLogger().info('Agent stopped');
+            callback();
+        }
+    );
 }
 
 exports.start = start;
